+++ im_library/client/im_requests/list_user_infrastructures.py
@@ -1,5 +1,3 @@
-from typing import Optional
-
 from im_library.client.base_requests import Get
 from im_library.client.query_parameters_base import QueryParametersBase
 from im_library.entities.enums.im_request_type import IMRequestType
